backend/migrations/requirements.txt
# Python dependencies for database migration scripts
psycopg2-binary==2.9.9
